Test Case/Fine-tuning/LLaMa Fine-tuning/4.QLoRA/requirements.txt
# QLoRA Fine-tuning Requirements
# 설치 방법: pip install -r requirements.txt

# PyTorch (CUDA 12.1 기준)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers 및 관련 라이브러리
transformers>=4.36.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.25.0

# QLoRA 핵심 라이브러리
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# 데이터 처리
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# 모니터링 및 로깅
wandb>=0.16.0
tensorboard>=2.15.0

# 환경 설정
python-dotenv>=1.0.0
PyYAML>=6.0

# GPU 모니터링
GPUtil>=1.4.0
psutil>=5.9.0

# Hugging Face
huggingface-hub>=0.19.0

# 기타 유틸리티
tqdm>=4.65.0
packaging>=23.0